[console] porting console tests to Jest (#19543)
* porting console tests to Jest

* adding missing import
