+++ package.json
@@ -275,6 +275,7 @@
     "istanbul-instrumenter-loader": "3.0.0",
     "jest": "^22.4.3",
     "jest-cli": "^22.4.3",
+    "jest-raw-loader": "^1.0.1",
     "jimp": "0.2.28",
     "jsdom": "9.9.1",
     "karma": "1.7.0",
