+++ src/core_plugins/console/public/src/__mocks__/ui/chrome.js
@@ -16,30 +16,3 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-import { getContentType } from '../../src/es';
-
-const { test, module, equal } = window.QUnit;
-
-const APPLICATION_JSON = 'application/json';
-
-module('Content type');
-
-test('body', () => {
-  const contentType = getContentType([
-    JSON.stringify({
-      foo: 'baz'
-    }),
-    JSON.stringify({
-      foo: 'bar'
-    })
-  ].join('\n'));
-
-  equal(contentType, APPLICATION_JSON);
-});
-
-test('no body', () => {
-  const contentType = getContentType('');
-
-  equal(contentType, undefined);
-});
