+++ src/core_plugins/console/public/src/output.js
@@ -18,7 +18,7 @@
  */
 
 import _ from 'lodash';
-const ace = require('ace');
+const ace = require('brace');
 const settings = require('./settings');
 const OutputMode = require('./sense_editor/mode/output');
 const smartResize = require('./smart_resize');
