+++ src/core_plugins/console/public/src/sense_editor/mode/input_highlight_rules.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-const ace = require('ace');
+const ace = require('brace');
 import { addToRules } from './x_json_highlight_rules';
 
 const oop = ace.acequire('ace/lib/oop');
