+++ src/core_plugins/console/public/src/sense_editor/mode/output.js
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-const ace = require('ace');
+const ace = require('brace');
 require('./output_highlight_rules');
