--- src/core_plugins/console/public/tests/index.html
@@ -1,36 +1,0 @@
-  <style type="text/css">
-    #editor_container {
-      display: none;
-      position: absolute;
-      top: 20px;
-      bottom: 20px;
-      left: 20px;
-      right: 20px;
-      z-index: 200;
-      border: 1px solid #333;
-    }
-
-    #output_container {
-      display: none;
-      position: absolute;
-      height: 250px;
-      width: 350px;
-      bottom: 20px;
-      right: 20px;
-      z-index: 201;
-      border: 1px solid #333;
-    }
-
-    #editor, #output {
-      height: 100%;
-      width: 100%;
-      position: relative;
-    }
-  </style>
-<div id="qunit"></div>
-<div id="editor_container">
-  <div id="editor"></div>
-</div>
-<div id="output_container">
-  <div id="output"></div>
-</div>
