+++ src/core_plugins/console/public/tests/src/content_type.test.js
@@ -17,4 +17,27 @@
  * under the License.
  */
 
-import './tests';
+import { getContentType } from '../../src/es';
+
+const APPLICATION_JSON = 'application/json';
+describe('Content type', () => {
+  test('body', () => {
+    const contentType = getContentType([
+      JSON.stringify({
+        foo: 'baz'
+      }),
+      JSON.stringify({
+        foo: 'bar'
+      })
+    ].join('\n'));
+
+    expect(contentType).toEqual(APPLICATION_JSON);
+  });
+
+  test('no body', () => {
+    const contentType = getContentType('');
+
+    expect(contentType).toBeUndefined();
+  });
+});
+
