+++ src/core_plugins/console/public/tests/src/curl_parsing.txt
