+++ src/core_plugins/console/public/tests/src/setup_mocks.js
@@ -16,35 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-const ace = require('brace');
-import 'brace/mode/json';
-import { addToRules } from './x_json_highlight_rules';
-
-const oop = ace.acequire('ace/lib/oop');
-const JsonHighlightRules = ace.acequire('ace/mode/json_highlight_rules').JsonHighlightRules;
-
-export function OutputJsonHighlightRules() {
-
-  this.$rules = {};
-
-  addToRules(this, 'start');
-
-  this.$rules.start.unshift(
-    {
-      'token': 'warning',
-      'regex': '#!.*$'
-    },
-    {
-      'token': 'comment',
-      'regex': '#.*$'
-    }
-  );
-
-  if (this.constructor === OutputJsonHighlightRules) {
-    this.normalizeRules();
-  }
-
-}
-
-oop.inherits(OutputJsonHighlightRules, JsonHighlightRules);
+/* eslint no-undef: 0 */
+jest.mock('../../src/sense_editor/mode/worker', () => { return { workerModule: { id: 'sense_editor/mode/worker', src: '' } }; });
+window.Worker = function () { this.postMessage = () => {}; this.terminate = () => {}; };
+window.URL = {
+  createObjectURL: () => { return ''; }
+};
+jest.mock('../../src/storage');
+document.queryCommandSupported = () => true;
