+++ src/dev/jest/config.js
@@ -80,6 +80,8 @@ export default {
   transform: {
     '^.+\\.js$': '<rootDir>/src/dev/jest/babel_transform.js',
     '^.+\\.tsx?$': '<rootDir>/src/dev/jest/ts_transform.js',
+    '^.+\\.txt?$': 'jest-raw-loader',
+    '^.+\\.html?$': 'jest-raw-loader',
   },
   transformIgnorePatterns: [
     '[/\\\\]node_modules[/\\\\].+\\.js$',
